requirements.txt
aiogram
python-dotenv
aiosqlite
sqlalchemy
pandas
openpyxl
